refactor: move resolveReportForMoneyRequest into IOUUtils
