+++ tests/unit/IOUUtilsTest.ts
@@ -35,6 +35,9 @@ function initCurrencyList() {
 jest.mock('@src/libs/Navigation/Navigation', () => ({
     navigate: jest.fn(),
     goBack: jest.fn(),
+    navigationRef: {
+        getCurrentRoute: jest.fn(() => undefined),
+    },
 }));
 
 describe('IOUUtils', () => {
@@ -867,6 +870,77 @@ describe('getExistingTransactionID', () => {
 
             expect(result1.chatReportID).toBeDefined();
             expect(result2.chatReportID).toBeDefined();
+        });
+    });
+
+    describe('resolveReportForMoneyRequest', () => {
+        const policyForResolve: Policy = {...createRandomPolicy(1, CONST.POLICY.TYPE.TEAM, 'Resolve Test Policy'), id: 'resolve-policy'};
+
+        const makeOutstandingReport = (reportID: string): Report => ({
+            ...createRandomReport(Number(reportID), undefined),
+            reportID,
+            policyID: policyForResolve.id,
+            type: CONST.REPORT.TYPE.EXPENSE,
+            stateNum: CONST.REPORT.STATE_NUM.OPEN,
+            statusNum: CONST.REPORT.STATUS_NUM.OPEN,
+        });
+
+        const makeRouteReport = (reportID: string): Report => ({
+            ...createRandomReport(Number(reportID), undefined),
+            reportID,
+            policyID: policyForResolve.id,
+            type: CONST.REPORT.TYPE.CHAT,
+        });
+
+        const makeTransaction = (reportID: string): Transaction => ({...createRandomTransaction(Number(reportID)), reportID});
+
+        it('returns undefined when the transaction is unreported', () => {
+            const transaction = makeTransaction(CONST.REPORT.UNREPORTED_REPORT_ID);
+            const transactionReport = makeOutstandingReport('500');
+            const routeReport = makeRouteReport('100');
+            expect(IOUUtils.resolveReportForMoneyRequest({transaction, transactionReport, routeReport, policy: policyForResolve})).toBeUndefined();
+        });
+
+        it('returns the picked report when it is outstanding (user-selected report wins)', () => {
+            const transaction = makeTransaction('500');
+            const transactionReport = makeOutstandingReport('500');
+            const routeReport = makeRouteReport('100');
+            expect(IOUUtils.resolveReportForMoneyRequest({transaction, transactionReport, routeReport, policy: policyForResolve})?.reportID).toBe('500');
+        });
+
+        it('returns undefined when the picked report is non-outstanding and differs from the route (forces a new optimistic IOU)', () => {
+            const transaction = makeTransaction('500');
+            const nonOutstandingPick: Report = {
+                ...makeOutstandingReport('500'),
+                policyID: 'someOtherPolicy',
+            };
+            const routeReport = makeRouteReport('100');
+            expect(IOUUtils.resolveReportForMoneyRequest({transaction, transactionReport: nonOutstandingPick, routeReport, policy: policyForResolve})).toBeUndefined();
+        });
+
+        it('returns the route report when no different transaction report has been picked', () => {
+            const transaction = makeTransaction('100');
+            const transactionReport = makeRouteReport('100');
+            const routeReport = makeRouteReport('100');
+            expect(IOUUtils.resolveReportForMoneyRequest({transaction, transactionReport, routeReport, policy: policyForResolve})?.reportID).toBe('100');
+        });
+
+        it('falls back to the transaction report when no route report exists (the !routeReport branch)', () => {
+            const transaction = makeTransaction('500');
+            const transactionReport = makeOutstandingReport('500');
+            expect(IOUUtils.resolveReportForMoneyRequest({transaction, transactionReport, routeReport: undefined, policy: policyForResolve})?.reportID).toBe('500');
+        });
+
+        it('returns undefined when the picked report is processing and policy harvesting is disabled', () => {
+            const transaction = makeTransaction('500');
+            const processingPick: Report = {
+                ...makeOutstandingReport('500'),
+                stateNum: CONST.REPORT.STATE_NUM.SUBMITTED,
+                statusNum: CONST.REPORT.STATUS_NUM.SUBMITTED,
+            };
+            const routeReport = makeRouteReport('100');
+            const harvestingDisabledPolicy: Policy = {...policyForResolve, harvesting: {enabled: false}};
+            expect(IOUUtils.resolveReportForMoneyRequest({transaction, transactionReport: processingPick, routeReport, policy: harvestingDisabledPolicy})).toBeUndefined();
         });
     });
 });
